refactor: Refatora o Controller do Missions para incluir o ResponseEntity nas minhas rotas e customizar a resposta das requisicoes do cliente

+++ src/main/java/dev/Java10x/CadastroDeNinjas/Missions/MissionsController.java
@@ -1,5 +1,8 @@
 package dev.Java10x.CadastroDeNinjas.Missions;
 
+import dev.Java10x.CadastroDeNinjas.Ninjas.NinjaDTO;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,31 +19,52 @@ public class MissionsController {
 
     // Post -- Mandar uma requisicao para criar as missoes
     @PostMapping("/create")
-    public MissionsDTO createMission(@RequestBody MissionsDTO mission) {
-        return missionsService.createMission(mission);
+    public ResponseEntity<String> createMission(@RequestBody MissionsDTO mission) {
+        MissionsDTO newMission = missionsService.createMission(mission);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Missao criada com sucesso: " + newMission.getName() + " (ID): " + newMission.getId());
     }
 
     // Get -- Mandar uma requisicao para mostrar as missoes
     @GetMapping("/list")
-    public List<MissionsDTO> listMissions() {
-        return missionsService.listMissions();
+    public ResponseEntity<List<MissionsDTO>> listMissions() {
+        List<MissionsDTO> missions = missionsService.listMissions();
+        return ResponseEntity.ok(missions);
     }
 
     @GetMapping("/list/{id}")
-    public MissionsDTO listMissionById(@PathVariable Long id) {
-        return missionsService.listMissionById(id);
+    public ResponseEntity<?> listMissionById(@PathVariable Long id) {
+        MissionsDTO mission = missionsService.listMissionById(id);
+        if (mission != null) {
+            return ResponseEntity.ok(mission);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Missao de ID: " + id + " nao encontrada.");
+        }
     }
 
     // Put -- Mandar uma requisicao para alterar as missoes
     @PutMapping("/update/{id}")
-    public MissionsDTO updateMission(@PathVariable Long id, @RequestBody MissionsDTO updatedMission) {
-        return missionsService.updateMission(id, updatedMission);
+    public ResponseEntity<?> updateMission(@PathVariable Long id, @RequestBody MissionsDTO updatedMission) {
+        MissionsDTO mission = missionsService.updateMission(id, updatedMission);
+        if (mission != null) {
+            return ResponseEntity.ok(mission);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Missao de ID: " + id + " nao encontrada.");
+        }
     }
 
     // delete -- Mandar uma requisicao para deletar as missoes
     @DeleteMapping("/delete/{id}")
-    public void deleteMissionById(@PathVariable Long id) {
-        missionsService.deleteMissionById(id);
+    public ResponseEntity<String> deleteMissionById(@PathVariable Long id) {
+        if (missionsService.listMissionById(id) != null) {
+            missionsService.deleteMissionById(id);
+            return ResponseEntity.ok("Missao com ID: " + id + " deletada com sucesso.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Missao com ID: " + id + " nao encontrada.");
+        }
     }
 
 }
